Empty String default to name value
Set the default value of the name field variable to an empty String (""). This is part of a conversion to make the empty String named Status unique among other empty String named Status objects. Null is no longer be allowed as a valid argument for the name value of a Status object.