Extra cases attached to useCase now have flags.
Added the stunBreak and deathless fields, as well as corresponding set methods. These values can toggle on and off additional cases added to the useCase. stunBreak, when false, says the useCase also checks that the owner of the skill is not stunned. deathless, when false, says the useCase also checks that the owner of the skill is not defeated.