Second set of javadoc changes to Status.
Moving most of the exposition about Status variables to the set methods of the StatusBuilder class. This moves the valuable comments to public members and away from private fields. In turn, private fields in Status and StatusBuilder, as well as getters in the Status will link to the StatusBuilder setters when appropriate in order to minimize duplication of said exposition.