Replaced methods for that set skill parameters.
Previously, one could adjust one value with the overloaded get and
modify methods using a given int value. These overloaded methods have
been replaced with ones that accept a varag of ints instead. How each of
these methods behave will be specific to the skills that override them,
with the default being a simple call to the parameterless version of the
method.