Expanded the logic of Squad.removeFighter().
Squads now check to see if they are in a battle before they either
remove of add fighters to themselves. Additional checks for null or non
existant fighters have been added to squad methods. These changes
cascaded to the Fighter and FgihterBuilder classes to allow them to set
team parameters to null. When this happens, the given null value is
always replaced with a unique Team object.