Added PrintLogger listener to skills.
Additionally, skills were changed so that tasks that were common to the
generation of all skills were filtered into one or two methods, such as
adding the PrintLogger listener to skills or calling build(). This also
forced the javadocs of each skill to hold the consquences of call
get(int...) as that method was no longer being overrid.